(feat: 게시판 조회수 표시, 글 삭제 시 confirm 추가,  UI 소폭 수정) (#18)
* (feat) 회원가입 페이지 라우팅

* (refactor: 게시물 조회 렌더링 방법 개선)

* (refactor: ResponseDTO에 조회수 필드 추가)

* (feat: 게시판 라우팅)

+++ src/components/board/BoardDetail.tsx
@@ -11,7 +11,6 @@ function BoardDetail(): JSX.Element {
   const { boardId } = useParams();
   const navigate = useNavigate();
 
-
   const fetchBoard = async () => {
     try {
       const response = await api.get(`http://localhost:7777/api/board/${boardId}`);
@@ -23,18 +22,23 @@ function BoardDetail(): JSX.Element {
 
   const handleDeleteSubmit = async (e) => {
     e.preventDefault();
-    const token = Cookies.get('key');
-    const headers = {
-      'Content-Type': 'application/json',
-      Authorization: `Bearer ${token}`,
-    };
-
-    const response = await api.delete(`http://localhost:7777/api/board/delete/${boardId}`, { headers });
-    if (response.data.status === 200) {
-      alert('게시글이 삭제되었습니다.');
-    } else {
-      alert(`${response.data.message}`);
-      navigate(-1); // 이전 페이지로 이동 (수정 폼 들어갔다가 나와서 목록으로 나가려는 사람들에 대한 예외처리 필요)
+    const shouldDelete = window.confirm('정말로 삭제하시겠습니까?'); 
+  
+    if (shouldDelete) {
+      const token = Cookies.get('key');
+      const headers = {
+        'Content-Type': 'application/json',
+        Authorization: `Bearer ${token}`,
+      };
+  
+      const response = await api.delete(`http://localhost:7777/api/board/delete/${boardId}`, { headers });
+      if (response.data.status === 200) {
+        alert('게시글이 삭제되었습니다.');
+        
+      } else {
+        alert(`${response.data.message}`);
+      }
+      navigate(-2);
     }
   };
 
@@ -44,8 +48,7 @@ function BoardDetail(): JSX.Element {
 
   useEffect(() => {
     fetchBoard();
-  }, [board]);
-
+  }, []);
 
   return (
     <div>
@@ -61,15 +64,34 @@ function BoardDetail(): JSX.Element {
             <TableBody>
               <TableRow>
                 <TableCell>
-                  <Typography sx={{display: 'flex', justifyContent: 'center'}}>[{board.category}] {board.title}   
-                    <Typography sx={{ color: 'grey' }}> {moment(board.moddate).format('YYYY-MM-DD HH:mm')}
-                    </Typography>
+                  <Typography sx={{ padding:1, color: 'grey', display: 'flex', justifyContent: 'left' }}> 
+                  {moment(board.moddate).format('YYYY-MM-DD HH:mm')}
                   </Typography>
-                  <Typography sx={{display: 'flex', justifyContent: 'center'}}> 작성자: {board.writerNickname}</Typography>
+                  <Typography sx={{display: 'flex', justifyContent: 'left', padding:1}}>[{(() => {
+                switch (board.category) {
+                  case 'walk-with':
+                    return '산책가요';
+                  case 'show-off':
+                    return '동물자랑';
+                  case 'sitter':
+                    return '시터공고';
+                  default:
+                    return board.category;
+                }
+              })()}] {board.title}  
+                  </Typography>
+                  <Typography sx={{display: 'flex', justifyContent: 'left', padding:1}}>
+                    👤 {board.writerNickname}
+                  </Typography>
                   <Divider sx={{margin:3}}/>
                   {board.image !== '' && <div style={{ display: 'flex', justifyContent: 'center', margin: '10px 0' }}>
                     <img src={board.image} alt={`${board.title}`} style={{ maxWidth: '100%' }} /></div>}
-                  <Typography sx={{display: 'flex', justifyContent: 'center', margin:5}}>{board.content}</Typography>
+                  <Typography sx={{display: 'flex', justifyContent: 'center', margin:5}}>
+                    {board.content}
+                  </Typography>
+                  <Typography sx={{display: 'flex', justifyContent: 'right', padding:1}}>
+                  👀 {board.clickCnt}
+                  </Typography>
                 </TableCell>
               </TableRow>
             </TableBody>
